+++ libs/pigment/colorspaces/KoRgbU8ColorSpace.cpp
@@ -71,40 +71,40 @@ KoColorSpace* KoRgbU8ColorSpace::clone() const
 
 void KoRgbU8ColorSpace::fromQColor(const QColor& c, quint8 *dst, const KoColorProfile * /*profile*/) const
 {
-    QVector<float> channelValues;
+    QVector<qreal> channelValues;
     channelValues << c.blueF() << c.greenF() << c.redF() << c.alphaF();
     fromNormalisedChannelsValue(dst, channelValues);
 }
 
 void KoRgbU8ColorSpace::toQColor(const quint8 * src, QColor *c, const KoColorProfile * /*profile*/) const
 {
-    QVector<float> channelValues(4);
+    QVector<qreal> channelValues(4);
     normalisedChannelsValue(src, channelValues);
     c->setRgbF(channelValues[2], channelValues[1], channelValues[0], channelValues[3]);
 }
 
-void KoRgbU8ColorSpace::toHSY(const QVector<double> &channelValues, qreal *hue, qreal *sat, qreal *luma) const
+void KoRgbU8ColorSpace::toHSY(const QVector<qreal> &channelValues, qreal *hue, qreal *sat, qreal *luma) const
 {
     
     RGBToHSY(channelValues[0],channelValues[1],channelValues[2], hue, sat, luma);
 }
 
-QVector <double> KoRgbU8ColorSpace::fromHSY(qreal *hue, qreal *sat, qreal *luma) const
+QVector <qreal> KoRgbU8ColorSpace::fromHSY(qreal *hue, qreal *sat, qreal *luma) const
 {
-    QVector <double> channelValues(4);
+    QVector <qreal> channelValues(4);
     HSYToRGB(*hue, *sat, *luma, &channelValues[0],&channelValues[1],&channelValues[2]);
     channelValues[3]=1.0;
     return channelValues;
 }
 
-void KoRgbU8ColorSpace::toYUV(const QVector<double> &channelValues, qreal *y, qreal *u, qreal *v) const
+void KoRgbU8ColorSpace::toYUV(const QVector<qreal> &channelValues, qreal *y, qreal *u, qreal *v) const
 {
     RGBToYUV(channelValues[0],channelValues[1],channelValues[2], y, u, v);
 }
 
-QVector <double> KoRgbU8ColorSpace::fromYUV(qreal *y, qreal *u, qreal *v) const
+QVector <qreal> KoRgbU8ColorSpace::fromYUV(qreal *y, qreal *u, qreal *v) const
 {
-    QVector <double> channelValues(4);
+    QVector <qreal> channelValues(4);
     YUVToRGB(*y, *u, *v, &channelValues[0],&channelValues[1],&channelValues[2]);
     channelValues[3]=1.0;
     return channelValues;
